% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_clear.R
\name{cron_clear}
\alias{cron_clear}
\title{Clear all cron jobs}
\usage{
cron_clear(ask = TRUE, user = "")
}
\arguments{
\item{ask}{Boolean; ask before removal?}

\item{user}{The user whose crontab we are clearing.}
}
\description{
Clear all cron jobs
}

