% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_load.R
\name{cron_load}
\alias{cron_load}
\title{Load a crontab from file}
\usage{
cron_load(file, user = "", ask = TRUE)
}
\arguments{
\item{file}{The file location of a crontab.}

\item{user}{The user for whom we will be loading a crontab.}

\item{ask}{Boolean; show prompt asking for validation}
}
\description{
Load a crontab from file
}
