% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_njobs.R
\name{cron_njobs}
\alias{cron_njobs}
\title{List the number of rCron cron jobs}
\usage{
cron_njobs(user = "")
}
\arguments{
\item{user}{The user whose cron jobs we wish to examine.}
}
\description{
List the number of rCron cron jobs
}

