% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_rstudioaddin.R
\name{cron_rstudioaddin}
\alias{cron_rstudioaddin}
\title{Launch an RStudio addin which allows to schedule an Rscript interactively.}
\usage{
cron_rstudioaddin(RscriptRepository = Sys.getenv("CRON_LIVE", unset = getwd()))
}
\arguments{
\item{RscriptRepository}{path to the folder where R scripts will be copied to and launched from, and by default log files will be written to.
Defaults to the current working directory or in case it is set, the path set in the \code{CRON_LIVE} environment variable.}
}
\value{
the return of \code{\link[shiny]{runGadget}}
}
\description{
Launch an RStudio addin which allows to schedule an Rscript interactively.
}
\examples{
\dontrun{
cron_rstudioaddin()
}
}
