% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{crond}
\alias{crond}
\title{Task Scheduler}
\usage{
crond(Schedules, TasksInfo, method = "callback",
  handler = "dummy_task_handler", ...)
}
\arguments{
\item{Schedules}{An ordered vector with the schedules for each task in shorthand period:offset form}

\item{TasksInfo}{Ordered Structure with extra information which will be passed to handler}

\item{method}{Whether a function (callback) or a program should be called}

\item{handler}{The name of the function which handles the task execution}

\item{...}{Extra arguments to be passed to the handler}
}
\description{
Run tasks in a schedule
}
\examples{
# Define a handler
my_handler <- function(task_id, timestamp, tasks_info) {
  # Do something with this "task_id" using the information in "tasks_info"
  message("Running ", task_id, " at ", unclass(Sys.time()), " - scheduled: ",
          timestamp, " (", as.POSIXct(timestamp, origin = "1970-01-01") , ")")
}
# Define the tasks info structure
my_tasks  <- data.frame(schedules = c("2s", "1m:3s", "1h:5s"),
                        my_jobs    = c("job01", "job02",  "job03"),
                        other_info = c("...",   "...",     "..."),
                        stringsAsFactors = F)
# create the schedules vector
schedules <- my_tasks$schedules
# Start the scheduler
crond(schedules, my_tasks, "callback", "my_handler")

}
