% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronr.R
\docType{package}
\name{cronr}
\alias{cronr}
\alias{cronr-package}
\title{cronr: A Cron-like second-precision Scheduler for R}
\description{
This is a scheduling package for running one-off or recurring tasks in second-precision schedules.
It may be useful either in maintenance tasks, or in batch processing pipelines where new data are made available periodically.
For example, you may need to retrieve new records written in a database every 30 seconds and process them in batches.
}
\section{crond functions}{

crond
}

