% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{dummy_task_handler}
\alias{dummy_task_handler}
\title{dummy_task_handler}
\usage{
dummy_task_handler(task_id, timestamp, tasks_info)
}
\description{
A dummy task handler that just prints an informative message. 
Arguments do not need to be supplied, they will passed by crond
}
